requirements.txt
pdfplumber>=0.9.0
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=1.0.0
librosa>=0.10.0
pydub>=0.25.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
yt-dlp>=2023.12.30
openai-whisper>=20231117